Move calendar driver to a new file
